Split auto rename planning modules
